[CHORE] NavigationUtil을 FragmentExt로 이름 변경
